Increase Coverage By Removing Code (#104)
In implementing #87, I realised that a lot of the missing code coverage
by lines were in the `UserOperationLib` source file. Since this is only
used by test contracts (in particular the test entry point and the mock
4337 Safe), I moved the code around/cleaned up dead code to increase the
coverage of the "main" contracts. In particular:

* The `UserOperation` struct definition moved to the 4337 interface file
(felt natural as the struct is part of the 4337 account interface)
* Moved the `UserOperationLib` library to the `test/` subdirectory, the
only place it was being used
* I also got rid of some dead code, library functions that weren't being
used anywhere
* Deduplicated the `Account` interface from the test entry point
* Fixed some Solidity warning lints (that weren't causing CI errors -
should we change CI to fail on these warnings as well?)