Increase branch Coverage (#108)
Partially fixes #87

This PR adds unit tests to cover the remaining branches from the
`SimpleEIP4337Module` contract. With this, we should have 100% coverage
for the core contracts in the repo (although, not all paths in the test
contracts are covered, which should be fine).

In terms of test organisation, I added a `EIP4337Module.spec.ts` for
unit tests directly targeting the module functions (`validateUserOp` and
`execUserOp*`).