Send encoded Safe operation bytes to `checkSignatures` call (#165)
Fixes #160

This PR fixes the `checkSignatures` call to the Safe to include both the
Safe operation hash for recovery, but also the encoded operation bytes,
which is required when using an ERC-1271 signer. In order to test this,
the `SafeMock` now requires that `dataHash == keccak256(data)` and I
added a unit test using a nested Safe set (note that this would require
a staked paymaster to work with 4337 bundlers, but the test does check
that the contract signatures works as expected).