[4337] User Operation with Custom Singleton Signers (#159)
Fixes #148 

This PR introduces an E2E test using the official bundler where we
deploy a Safe that uses a custom (and silly 🤪) signing scheme over
ERC-1271. Specifically, the signing scheme is deployed as a singleton
contract that can be shared by multiple Safes using a `mapping(safe =>
data)` to differentiate between signers. In theory, this could be
adapted to, for example, checking a signature using P256 curve in order
to integrate with WebAuthn, and is meant to “prove” that the Safe with
the ERC-4337 module would work with existing WebAuthn signer
implementations.

The actual implementation is a little convoluted for a few reasons.
Namely, ERC-4337 only allows a **single** `CREATE2` per user operation
that **MUST** result in the `userOp.sender` account. Because of this,
deploying ad-hoc signers in the `Safe.setup` call is not possible, and
we have instead opted for a “singleton signer” pattern, where the logic
for verifying signatures is shared and a mapping is used to store
per-Safe context relevant to signature verification.

Additionally, In order for this to work with ERC-4337, a staked factory
is needed (as accessing associated storage for user operations with
`initCode` require a factory). Unfortunately, staked factories add all
kinds of complexities:
- Currently, the staked factory contract is quite dumb (can only stake,
but can’t unstake). A full staked factory contract requires us to iron
out some requirements and make a proper implementation with
documentation and tests (out of scope for this PR and issue, and a
non-trivial amount of work)
- Staking is determined per bundler, so while some amount of stake may
be sufficient for a bundler, it may not be sufficient for all (which
adds additional off-chain management dependencies).

**For these reasons, I would recommend exploring #149**.