Add E2E Test with WebAuthn Signer (#197)
A huge part of the way to #173

This PR adds an E2E test using the reference 4337 EntryPoint and bundler
for a Safe deployment with a WebAuthn (i.e. Passkeys) signer.

Note that this E2E test is not 100% complete. In particular, we don't
get test using the Safe after the launchpad deployment is completed. I
want to add that test in a follow-up when I actually add the gas
profiling for the WebAuthn signer.