Add an updated EIP-1271 function to the WebAuthn signer (#205)
This PR:
- Adds an updated EIP-1271 validation method to the WebAuthnSigner
contract for futureproofing
- I created an abstract contract with both magic values, similar to how
it's done in the Safe Account contract. We cannot reuse the Safe's
contract because it only includes one value (either the legacy or the
new one)
- I had to adjust the return values of signature-validating functions
slightly. For the generic `checkSignatures` function, I changed it to a
boolean, similar to how it's done in the underlying FCL's validation
function