New Signature Encoding for Webauthn Signer (#221)
Fixes #220 

This PR proposes a new signature encoding scheme for the Webauthn
signer. This takes advantage of the [`clientDataJSON` serialization
specification]() which requires a specific ordering and formatting (no
additional whitespace) for the fields). Notably, this scheme:

* reduces the amount of calldata by 114 bytes per signature; this is
non-trivial in the context of reducing calldata for rollups (where
calldata is more expensive than computation).
* no longer requires a trusted `challengeOffset` (which has some
security implications).

I plan to upstream these changes to FCL after doing a bit of
investigation on potentially optimizing:
- base64 encoding of the challenge (since we use a fixed 32-byte
challenge while the library allows for arbitrary `bytes`)
- byte concatenation logic to try and make it slightly more gas
efficient