Fix E2E Tests For ERC-4337 v0.7 Bundler (#300)
Fixes #227 

This PR updates the E2E tests to work with the latest bundler for the
EntryPoint contract v0.7. Note that there was a notable change in that
the `PackedUserOperation` struct is only used on-chain, and the bundler
RPC uses a different `UserOperation` type that is completely unpacked.
This required some additional helper methods as well as some adjustments
to tests to distinguish between the packed and unpacked versions of user
operations.

Also, unfortunately the reference bundler repository is not tagged yet -
so we are using a commit hash that corresponds to the latest version. I
asked the 4337 team to create a new tag, at which point, I will update
the docker file. Additionally, the latest bundler version uses
submodules, so some dockerfile tweaking was required.

Note that the broken CI is expected until #270 is fixed.