Feature: Implement a Passkey `SignatureValidator` factory (#306)
This PR is a partial solution for
https://github.com/safe-global/safe-modules/issues/288. What needs to be
added is a proper test suite for the factory. It only has one migrated
end-to-end test from the 4337 module package.

What was done:
- Introduce `ICustomECSignerFactory` and `ICustom256BitECSignerFactory`
interfaces for custom signer factories. The first is generic and should
work with signing algorithms of any key length, while the second is
optimized for EC signing and 256-bit key sizes. The first one is not
used in the project and is kept as an example or for future reference.
- Add the `WebAuthnVerifier` and make it use the new P256Library
introduced in #298
- Migrate the `SafeSignerLaunchpad` contract and make it implement the
`ICustom256BitECSignerFactory` interface thus making it
`Safe256BitECSignerLaunchpad`
- Migrate and adjust the end-to-end test from the 4337 module for the
launchpad

To be done:
- Remove the migrated contracts from the 4337 package