Verify Multiple P-256 Verifiers Are Supported (#307)
Partially addresses #285

This PR adds new tests that verify multiple P-256 verifiers are
supported by our code. The `webauthn` shim code now moved to the
`passkey` package and is imported by the `4337` package, as this made
more sense to me. @mmv08 moved E2E tests for Passkey+4337 to the
`passkey` project, so once they both merge, the dependency can be
removed.

Note that for now, we use a `TestWebAuthnSignerFactor` contract just for
implementing the tests, but it should be switched to using the canonical
signing factory from #306 once merged.

The additional verifier that is being used is the one from
[Daimo-eth](https://github.com/daimo-eth/p256-verifier). The artifact
was vendored into the project, as it is a Foundry project, and this is
the easiest way to include the dependency without building it (giving us
identical bytecode to what is deployed on-chain).

Adding a test for using the precompile will be done in a separate PR.

---

Note that this PR contains some unrelated changes to the deployment and
E2E setup. This was a result of rebasing onto changes introduced in #306
to get things working as well as some nits leftover from the
aforementioned PR.