Move 4337 Testing Utilies To Local Bundler Package (#315)
This PR moves the TS testing utility code for 4337 local bundler tests
into the package introduced in #314. This is done in an effort to avoid
duplicating the `test/utils/e2e.ts` file that contains setup functions
for the local bundler setup that gets deployed from the
`safe-4337-local-bundler` package.