Refactor WebAuthn Implementation (#320)
This PR refactors the WebAuthn implementation into a library instead of
split into multiple contracts. This removes the amount of `CALL`s
required for verifying a signature.

I did some analysis on the deployment vs. signature verification costs
comparing this code, `main` and this change incorporating optimizations
from #289 and found this:

| setup | deployment | verification (dummy) | break even |
| ------------------------------------ | ---------- |
-------------------- | ---------- |
| main (viaIR = false) | 373543 | 17789 | 0 |
| main (viaIR = true) | 393494 | 16443 | 15 |
| this branch (viaIR = false) | 612123 | 13835 | 61 |
| this branch (viaIR = true) | 594058 | 12551 | 43 |
| base64 optimizations (viaIR = false) | 442545 | 12965 | 15 |
| base64 optimizations (viaIR = true) | 414941 | 11102 | 7 |

Meaning that after 7 signatures, even with the slightly larger
deployment costs, we would break even by avoiding the extra `CALL`
(noting that ABI-encoding the call to the `WebAuthnVerifier` is not
particularly efficient in the first place).