Add Gas Benchmarking Tests for WebAuthn Signer (#324)
This PR adds gas benchmarking tests for the WebAuthn signer. Note that
we got rid of the `MultipleVerifiers.spec.ts` file, as it is essentially
the new gas benchmarking tests, and there was no need to duplicate the
work.

The benchmarks look like this:
```
  Gas Benchmarking
    WebAuthnSigner
      ⛽ deployment: 612123
      ✔ Benchmark signer deployment cost (694ms)
      ⛽ verification (FreshCryptoLib): 219365
      ✔ Benchmark signer verification cost with FreshCryptoLib verifier (172ms)
      ⛽ verification (daimo-eth): 351273
      ✔ Benchmark signer verification cost with daimo-eth verifier (201ms)
      ⛽ verification (Dummy): 13835
      ✔ Benchmark signer verification cost with Dummy verifier
```

Note that we include a "dummy" benchmark, this is because gas
consumption of P-256 signature verification is unstable, and this allows
us to better compare gas characteristics of the `WebAuthn` signing
message computation overhead.

Furthermore, we remove the check on whether or not the P-256 verifier
has code. This is important as precompiles have no code, so our WebAuthn
stuff wouldn't work with precompiles without this change.