Fix P-256 Test Account Implementation (#338)
This PR fixes a small bug in the P-256 account implementation used for
tests. It provides an interface that exposes both a high and low S value
for a P-256 ECDSA signature, but computes it in a way that the ordering
isn't stable (and leads to failed assertions some of the times).

Likely a copy-pasta issue from copying signing code from the
`webauthn.ts` module (which does not care abot signature malleability).