Use assembly in WebAuthnVerifier for Base64 encoding (#318)
This PR implements #289 

The algorithm is implemented as a library instead of the suggested
in-place adjustment from the original issue. Akshay's original
implementation was doing it in-place (which can still be found in the
commit history). While it was efficient code size-wise, but gas-wise, it
consumed more gas than FCL's implementation, probably due to some
inefficient stack manipulations. I tested it with Solady's
[library](https://github.com/vectorized/solady/blob/main/src/utils/Base64.sol)
then, and it turned out to be much more efficient even though it wasn't
doing the encoding in place.

The implementation I'm proposing is based on Solady's efficient
algorithm. It seems to me that it's not adjustable to in-place encoding
due to the fact the algorithm takes advantage of optimizing stack
operations by writing X bytes to the scratch space and then writing them
to the string at position Y+(I*N) (I=Iteration, Y=start, N=bytes written
in an iteration) as a 32-byte word. Because it writes them as a 32-byte
word, it’s not suitable for modifying a string in place because it will
overwrite the existing string bytes with empty characters. (I'm happy if
I'm wrong though)

**Changes in PR**:
- Add `codesize` task
- Add a `Base64Url.sol` library with a Base64Url encoding algorithm that
is optimized for encoding `bytes32` data.
- Add tests for the base64 library
- Edit the typescript configuration to include the hardhat.config.ts so
all the type extensions are picked up. Necessary to get the project
compiled.

### Code size change

This PR:
`WebAuthnSigner 2693 bytes (limit is 24576)`

Default main branch:
`WebAuthnSigner 2817 bytes (limit is 24576)`

### Gas Usage - Needs additional test for benchmarking

**Solady**:
```
  Gas Benchmarking
    WebAuthnSigner
      ⛽ deployment: 550077
      ✔ Benchmark signer deployment cost
      ⛽ verification (FreshCryptoLib): 216563
      ✔ Benchmark signer verification cost with FreshCryptoLib verifier (70ms)
      ⛽ verification (daimo-eth): 347246
      ✔ Benchmark signer verification cost with daimo-eth verifier (116ms)
      ⛽ verification (Dummy): 12713
      ✔ Benchmark signer verification cost with Dummy verifier
```

**FCL**:
```
  Gas Benchmarking
    WebAuthnSigner
      ⛽ deployment: 612929
      ✔ Benchmark signer deployment cost
      ⛽ verification (FreshCryptoLib): 217272
      ✔ Benchmark signer verification cost with FreshCryptoLib verifier (70ms)
      ⛽ verification (daimo-eth): 347636
      ✔ Benchmark signer verification cost with daimo-eth verifier (108ms)
      ⛽ verification (Dummy): 13844
      ✔ Benchmark signer verification cost with Dummy verifier
```

**This PR**:
```
  Gas Benchmarking
    WebAuthnSigner
      ⛽ deployment: 533456
      ✔ Benchmark signer deployment cost
      ⛽ verification (FreshCryptoLib): 210645
      ✔ Benchmark signer verification cost with FreshCryptoLib verifier (64ms)
      ⛽ verification (daimo-eth): 337941
      ✔ Benchmark signer verification cost with daimo-eth verifier (108ms)
      ⛽ verification (Dummy): 12281
      ✔ Benchmark signer verification cost with Dummy verifier
```

**This PR (viaIR)**:
```
  Gas Benchmarking
    WebAuthnSigner
      ⛽ deployment: 544844
      ✔ Benchmark signer deployment cost
      ⛽ verification (FreshCryptoLib): 294440
      ✔ Benchmark signer verification cost with FreshCryptoLib verifier (108ms)
      ⛽ verification (daimo-eth): 335453
      ✔ Benchmark signer verification cost with daimo-eth verifier (117ms)
      ⛽ verification (Dummy): 10352
      ✔ Benchmark signer verification cost with Dummy verifier
```

---------

Co-authored-by: Mikhail Mikheev <16622558+mmv08@users.noreply.github.com>
Co-authored-by: Nicholas Rodrigues Lordello <nick@safe.global>