Base-64 Encode Challenge Directly Into JSON String
This PR encodes the 32 byte challenge directly into the client data JSON
that needs to be computed for the signing message. This allows the code
to avoid copy loops when computing the client data JSON that would be
required pre-`mcopy` opcode if the challenge were encoded to a separate
base-64 string.

This change has some nice small improvements to the overall cost of the
verifying signatures, and more substantial improvements to the
deployment cost of the signer.