Generalize Signing Message Encoding (#362)
Closes #292 

This PR generalizes the `WebAuthn` library to work better with future
public key algorithms that we may want to implement. Notably, the
hashing algorithm used for signing depends on the public key algorithm
of the credential, and is not always SHA-256. In contrast, the client
data JSON is **always** hashed with SHA-256. As such, we refactored the
function to return the actual signing message before hashing, so it can
be more easily composed in the eventuality that we start to support
other algorithms.

We also refactor encoding to be done in assembly for some small gas
wins, because Solidity is double copying the authenticator data.

<details><summary>Before:</summary>

```
  Gas Benchmarking [@bench]
    WebAuthnSigner
      ⛽ deployment: 471398
      ✔ Benchmark signer deployment cost (761ms)
      ⛽ verification (Dummy): 11148
      ✔ Benchmark signer verification cost with Dummy verifier
```

</details>

<details><summary>After:</summary>

```
  Gas Benchmarking [@bench]
    WebAuthnSigner
      ⛽ deployment: 468188
      ✔ Benchmark signer deployment cost (744ms)
      ⛽ verification (Dummy): 11126
      ✔ Benchmark signer verification cost with Dummy verifier
```

</details>

A more significant change is the re-ordering of operations in the
`verifySignature` function. This is a workaround to
https://github.com/ethereum/solidity/issues/14934 as it "tricks" the Yul
IR optimizer to perform much better (especially WRT code size) when Yul
optimizations are enabled:

<details><summary>Before (with Yul IR optimizations):</summary>

```
  Gas Benchmarking [@bench]
    WebAuthnSigner
      ⛽ deployment: 515182
      ✔ Benchmark signer deployment cost (742ms)
      ⛽ verification (Dummy): 9958
      ✔ Benchmark signer verification cost with Dummy verifier
```

</details>

<details><summary>After (with Yul IR optimizations):</summary>

```
  Gas Benchmarking [@bench]
    WebAuthnSigner
      ⛽ deployment: 470475
      ✔ Benchmark signer deployment cost (727ms)
      ⛽ verification (Dummy): 9624
      ✔ Benchmark signer verification cost with Dummy verifier
```

</details>