Full Coverage of `SafeWebAuthnSignerFactory` Contract (#373)
Fixes #365 

This PR adds missing branch coverage for the `SafeWebAuthnSignerFactory`
contract so that it is fully covered.

Note that we changed a `require` to an `assert`. This was prompted by
the Solidity coverage tool complaining that all paths weren't taken.
However, having that path not be taken would imply that the address
derivation for `CREATE2` to not function as expected, so should never
happen on EVM chains (note to self, do we require a slightly different
contract for zkSync?). Looking through the Solidity documentation, I
found that `assert` was more applicable here, and luckily the coverage
tool does not expect the negative path to be explored. From the docs:

> Assert should only be used to test for internal errors, and to check
invariants. Properly functioning code should never create a Panic, not
even on invalid external input. If this happens, then there is a bug in
your contract which you should fix. Language analysis tools can evaluate
your contract to identify the conditions and function calls which will
cause a Panic.

According to this description, this is a exemplar use of the `assert`
builtin.