Passkey example application: A Major refactoring (#378)
This PR lays the foundation for
https://github.com/safe-global/safe-modules/issues/360 and
https://github.com/safe-global/safe-modules/issues/361.

**Motivation**
The example application was built as a mere example, and all of the
state/screen management was done in a single file—` src/App.ts`—through
tens of different boolean and state variables. Since we have now decided
to add more features to the app, the approach is no longer scalable.

**Solution**
Isolate every step in the deployment flow - in other words, make every
step in a process its route. This should also help the developers to
comprehend the example since now, to understand what is needed to create
a passkey, they only have to review the `CreatePasskey` route, not mixed
with anything else.

**This PR**
- Add `react-router-dom` to manage the routing in the application
- Create a route for every step in the process
- Introduce a new Safe Dashboard page with a link to the Safe{Wallet}
App (This page will be used for #360, the new action will be added here)

**Demo**
https://www.loom.com/share/b5dcec589cc44fd8be8cefe1ef9cb306

---------

Co-authored-by: Aaron Cook <aaron@safe.global>